fix(rpc): cast bigint args in is_garbage_time calls
get_home_away_splits, get_conference_splits, and get_red_zone_splits call
public.is_garbage_time(p.period, p.score_diff) directly against core.plays,
where period/score_diff are bigint -- Postgres does not implicitly downcast
for function resolution, so all three RPCs have errored with 'function
public.is_garbage_time(bigint, integer) does not exist' since the Tier 1
garbage-time change deployed. (down_distance/field_position were unaffected:
they read the precomputed marts.play_epa.is_garbage_time boolean.)

Surfaced by the MCP server's situational_splits tool; also breaks cfb-app's
Home/Away, Conference, and Red Zone team-page tabs in production.

+++ src/schemas/public/006_play_analysis_functions.sql
@@ -191,7 +191,7 @@ BEGIN
         WHERE g.season = p_season
           AND (p.offense = p_team OR p.defense = p_team)
           AND p.yardline >= 80
-          AND NOT public.is_garbage_time(p.period, p.score_diff)
+          AND NOT public.is_garbage_time(p.period::integer, p.score_diff::integer)
     )
     SELECT
         rzd.side::TEXT,
